add: Fruit Enum 객체 분리 차 파일 생성, Assets, AppDelegate, SceneDelegate 폴더 생성 및 해당 파일 분리
