+++ JuiceMaker/JuiceMaker/Model/Fruit.swift
@@ -1,0 +1,16 @@
+//
+//  Fruit.swift
+//  JuiceMaker
+//
+//  Created by 김민제 on 1/25/24.
+//
+
+import Foundation
+
+enum Fruit: String {
+    case strawberry = "딸기"
+    case banana = "바나나"
+    case pineapple = "파인애플"
+    case kiwi = "키위"
+    case mango = "망고"
+}
