+++ JuiceMaker/JuiceMaker/Model/FruitStore.swift
@@ -6,14 +6,6 @@
 
 import Foundation
 
-enum Fruit: String {
-    case strawberry = "딸기"
-    case banana = "바나나"
-    case pineapple = "파인애플"
-    case kiwi = "키위"
-    case mango = "망고"
-}
-
 class FruitStore {
     var fruitStorage: [Fruit: Int]
